// src/types.ts
export interface UnebPassSlip {
  indexNo: string;
  name: string;
  year: string;
  level: 'P.L.E' | 'U.C.E' | 'U.A.C.E';
  subjects: { name: string; grade: string }[];
  dateOfBirth?: string;
  schoolName?: string;
  schoolAddress?: string;
  entryCode?: string;
  aggregate?: string;
  result?: string;
}

export interface UnebStats {
    totalSlips: number;
    uniqueSchools: number;
    byLevel: {
        'P.L.E': { studentCount: number; years: string[] };
        'U.C.E': { studentCount: number; years: string[] };
        'U.A.C.E': { studentCount: number; years: string[] };
    };
}